+++ src/main/java/net/geforcemods/securitycraft/blocks/DisguisableBlock.java
@@ -7,6 +7,7 @@ import net.geforcemods.securitycraft.compat.IOverlayDisplay;
 import net.geforcemods.securitycraft.items.ModuleItem;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.NbtUtils;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -14,6 +15,7 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.HitResult;
@@ -106,10 +108,16 @@ public abstract class DisguisableBlock extends OwnableBlock implements IOverlayD
 			ItemStack module = be.hasModule(ModuleType.DISGUISE) ? be.getModule(ModuleType.DISGUISE) : ItemStack.EMPTY;
 
 			if (!module.isEmpty()) {
-				Block block = ((ModuleItem) module.getItem()).getBlockAddon(module.getTag());
+				BlockState disguisedState = NbtUtils.readBlockState(module.getOrCreateTag().getCompound("SavedState"));
 
-				if (block != null)
-					return Optional.of(block.defaultBlockState());
+				if (disguisedState != null && disguisedState.getBlock() != Blocks.AIR)
+					return Optional.of(disguisedState);
+				else { //fallback, mainly for upgrading old worlds from before the state selector existed
+					Block block = ((ModuleItem) module.getItem()).getBlockAddon(module.getTag());
+
+					if (block != null)
+						return Optional.of(block.defaultBlockState());
+				}
 			}
 		}