+++ src/main/java/net/geforcemods/securitycraft/models/DisguisableDynamicBakedModel.java
@@ -14,38 +14,35 @@ import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.client.model.data.IDynamicBakedModel;
 import net.minecraftforge.client.model.data.IModelData;
 import net.minecraftforge.client.model.data.ModelProperty;
-import net.minecraftforge.registries.ForgeRegistries;
 
 public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
-	public static final ModelProperty<ResourceLocation> DISGUISED_BLOCK_RL = new ModelProperty<>();
-	private final ResourceLocation defaultStateRl;
+	public static final ModelProperty<BlockState> DISGUISED_STATE_RL = new ModelProperty<>();
 	private final BakedModel oldModel;
 
-	public DisguisableDynamicBakedModel(ResourceLocation defaultStateRl, BakedModel oldModel) {
-		this.defaultStateRl = defaultStateRl;
+	public DisguisableDynamicBakedModel(BakedModel oldModel) {
 		this.oldModel = oldModel;
 	}
 
 	@Override
 	public List<BakedQuad> getQuads(BlockState state, Direction side, Random rand, IModelData modelData) {
-		ResourceLocation rl = modelData.getData(DISGUISED_BLOCK_RL);
+		BlockState disguisedState = modelData.getData(DISGUISED_STATE_RL);
 
-		if (rl != null && rl != defaultStateRl) {
-			Block block = ForgeRegistries.BLOCKS.getValue(rl);
+		if (disguisedState != null) {
+			Block block = disguisedState.getBlock();
 
-			if (block != null) {
-				final BakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(block.defaultBlockState());
+			if (block != Blocks.AIR) {
+				final BakedModel model = Minecraft.getInstance().getBlockRenderer().getBlockModel(disguisedState);
 
 				if (model != null && model != this)
-					return model.getQuads(block.defaultBlockState(), side, rand, modelData);
+					return model.getQuads(disguisedState, side, rand, modelData);
 			}
 		}
 
@@ -54,13 +51,13 @@ public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
 
 	@Override
 	public TextureAtlasSprite getParticleIcon(IModelData modelData) {
-		ResourceLocation rl = modelData.getData(DISGUISED_BLOCK_RL);
+		BlockState state = modelData.getData(DISGUISED_STATE_RL);
 
-		if (rl != null && rl != defaultStateRl) {
-			Block block = ForgeRegistries.BLOCKS.getValue(rl);
+		if (state != null) {
+			Block block = state.getBlock();
 
-			if (block != null && !(block instanceof DisguisableBlock))
-				return Minecraft.getInstance().getBlockRenderer().getBlockModel(block.defaultBlockState()).getParticleIcon(modelData);
+			if (block != Blocks.AIR)
+				return Minecraft.getInstance().getBlockRenderer().getBlockModel(state).getParticleIcon(modelData);
 		}
 
 		return oldModel.getParticleIcon(modelData);
@@ -76,13 +73,13 @@ public class DisguisableDynamicBakedModel implements IDynamicBakedModel {
 				Optional<BlockState> disguisedState = disguisedBlock.getDisguisedBlockState(level, pos);
 
 				if (disguisedState.isPresent()) {
-					tileData.setData(DISGUISED_BLOCK_RL, disguisedState.get().getBlock().getRegistryName());
+					tileData.setData(DISGUISED_STATE_RL, disguisedState.get());
 					return tileData;
 				}
 			}
 		}
 
-		tileData.setData(DISGUISED_BLOCK_RL, defaultStateRl);
+		tileData.setData(DISGUISED_STATE_RL, Blocks.AIR.defaultBlockState());
 		return tileData;
 	}