+++ src/main/java/net/geforcemods/securitycraft/network/server/SyncProjector.java
@@ -4,6 +4,7 @@ import java.util.function.Supplier;
 
 import net.geforcemods.securitycraft.blockentities.ProjectorBlockEntity;
 import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.NbtUtils;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
@@ -13,6 +14,7 @@ import net.minecraftforge.fmllegacy.network.NetworkEvent;
 public class SyncProjector {
 	private BlockPos pos;
 	private int data;
+	private BlockState state;
 	private DataType dataType;
 
 	public SyncProjector() {}
@@ -23,12 +25,20 @@ public class SyncProjector {
 		this.dataType = dataType;
 	}
 
+	public SyncProjector(BlockPos pos, BlockState state) {
+		this.pos = pos;
+		this.state = state;
+		this.dataType = DataType.BLOCK_STATE;
+	}
+
 	public static void encode(SyncProjector message, FriendlyByteBuf buf) {
 		buf.writeBlockPos(message.pos);
 		buf.writeEnum(message.dataType);
 
 		if (message.dataType == DataType.HORIZONTAL)
 			buf.writeBoolean(message.data == 1);
+		else if (message.dataType == DataType.BLOCK_STATE)
+			buf.writeNbt(NbtUtils.writeBlockState(message.state));
 		else
 			buf.writeVarInt(message.data);
 	}
@@ -41,6 +51,8 @@ public class SyncProjector {
 
 		if (message.dataType == DataType.HORIZONTAL)
 			message.data = buf.readBoolean() ? 1 : 0;
+		else if (message.dataType == DataType.BLOCK_STATE)
+			message.state = NbtUtils.readBlockState(buf.readNbt());
 		else
 			message.data = buf.readVarInt();
 
@@ -72,6 +84,9 @@ public class SyncProjector {
 					case HORIZONTAL:
 						be.setHorizontal(message.data == 1);
 						break;
+					case BLOCK_STATE:
+						be.setProjectedState(message.state);
+						break;
 					case INVALID:
 						break;
 				}
@@ -89,6 +104,7 @@ public class SyncProjector {
 		RANGE,
 		OFFSET,
 		HORIZONTAL,
+		BLOCK_STATE,
 		INVALID;
 	}
 }